refactor: make submission impl non blocking
