+++ Cargo.toml
@@ -30,6 +30,7 @@ secp256k1 = { version = "0.28.0", features = ["rand-std"] }
 serde = "1.0.188"
 serde_json = "1.0.107"
 serde_with = "3.3.0"
+tokio = "1.35"
 uuid = "1.6.1"
 
 [dev-dependencies]
